% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_step_metrics_multiple.R
\name{get_step_metrics_multiple}
\alias{get_step_metrics_multiple}
\title{Movement Metric Calculator}
\usage{
get_step_metrics_multiple(
  movement_data,
  id = "ID",
  x = "X",
  y = "Y",
  step_number = NA
)
}
\arguments{
\item{movement_data}{a data_frame with at least 3 columns containing individual identifiers, x, and y coordinate data.}

\item{id}{character giving name of column with unique identifiers}

\item{x}{character giving name of column with x coordinates}

\item{y}{haracter giving name of column with y coordinates}

\item{step_number}{name of a column with the ordered number of steps. Used to ensure proper sorting}
}
\value{
data frame
}
\description{
calculates distance traveled between all consecutive points (rows) in the data frame, with breaks by individual identifiers
}
\examples{
move_data1 <- data.frame(x = rnorm(10), y = rnorm(10), Individual = "Subject1", step = 1:10)
move_data2 <- data.frame(x = rnorm(10), y = rnorm(10), Individual = "Subject2", step = 1:10)
get_step_metrics_multiple(rbind(move_data1, move_data2), ID = "Individual", x = "x", y = "y", step_number = "step"
}
