% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify_by_group.R
\name{modify_by_group}
\alias{modify_by_group}
\title{Modify by Group}
\usage{
modify_by_group(raw_data, groups, FUNC)
}
\arguments{
\item{raw_data}{a dataframe}

\item{groups}{a vector of character strings giving names of columns to group by. Use NA if raw_data is already grouped}

\item{FUNC}{a function that takes a dataframe is input or a formula that can be converted to such a function.}
}
\value{
a dataframe, form determined by the function in FUNC
}
\description{
modifies a dataframe by applying a transformative function accross groups
}
\examples{
\dontrun{ 
sean_tracks2 <- read.csv("//10.126.19.90/EAS_ind/avining/data/Misc/sean_tracks2.csv")
sean_tracks3 <- modify_by_group(sean_tracks2, c("track_name_old"), ~mutate(., waypoint_number = seq_along(timestamp2)))
FFT_subset <- read.csv("//10.126.19.90/EAS_ind/avining/R Workshops/Session 3/FFT_subset.csv")
FFT_subset <- modify_by_group(FFT_subset, c("individual.local.identifier", "day"), ~get_step_metrics_individual(., x = "location.long", y = "location.lat"))
}
}
